PrefabCollection and Instantiation
Moved the prefabs into a PrefabCollection object, which lives in the
scene. This makes it accessible outside of just editor mode (during
run-time).

Fixed some bugs with correct sorting order from the SceneRenderer /
CameraViewRenderer.

Started working on Instantiation, as part of what the Map does. Still
thinking about  this.

Vastly updated the UML diagram to reflect the editor and game sides.
Shows the clear separation and one-directional flow of information.