Major Refactoring Part 1
The GameEditor has been refactored into four sections:

1. Sprites
2. Prefabs
3. Scene
4. Map

Each of these sections is now a manager, and it delegates all access to
its related classes. For example, if you want to access the SpritesUI,
you have to ask the SpriteManager for that.

My hope with this change is a clear dependency diagram that is easy to
follow, and easy to expand on.

Part two just involves more refactoring.